Removed most of the unused code.
